Move regxml_extractor.sh to autotool-generating form
OS X Snow Leopard's default Python is too old for the DFXML library.
Autotools can select an appropriate version of Python, and also handle
the installed-script location issue better than checking for the current
script path.

regxml_extractor.sh now has a ./configure-based required step.

On Snow Leopard, @bindir@ defaulted to '/bin', so an overriding bindir
is defined.  That default was contrary to the value we expected from
this page:
<http://www.gnu.org/prep/standards/html_node/Directory-Variables.html>